feat: interactive `world create` (#3)
Closes: WORLD-389

## What is the purpose of the change
<img src="https://github.com/Argus-Labs/world-cli/assets/4402697/8eede2a0-e314-4309-bbf2-4395bc852065" width=500><br/>

This change adds an interactive `world create` command. When used without any argument, it will start an interactive prompt requesting a game shard name from the user. Alternatively, if an argument is passed in, it will use it as the game shard name.

When `world create` is run, it will clone `starter-game-template` to a directory based on the name passed in by user. After the clone is successful, it will re-initialize git to remove the git history of starter-game-template.

### Changelog
- Add interactive `world create`
- Add basic style library that contains icon, text styling, and container style
- Create a `steps` component that makes it easier to do interactive step-by-step TUI
- Update license to BUSL
- Cleanup cobra's root.go
- Add PR template

## Testing and Verifying

1. Run `world create` (without arguments)
2. Run `world create <arg>`