fix: check git configuration before exec world create (#74)
Closes: WORLD-1177

## Overview

When user trying to exec `world create` but git configuration `.gitconfig` for username and password is not configured it will be returned error, but the gameshard dir is already created.

The issue is because on the last step of `world create` it will exec git commit, and it will need username and email to be configured in `.gitconfig`

Solutions
Before executing `world create`, world cli will check the git configuration is configured or not. if it's not, world cli will tell the command how to configure the git config.

## Brief Changelog

- Create func for checking git configuration
- Call the function at the first execution of `world create` command

## Testing and Verifying
- Added unit test for checkGitConfig func